refactor: move collection orgs inside grouped dir
