+++ src/components/organisms/Collections/CollectionHero/CollectionHero.types.ts
