+++ src/components/templates/Collection/Collection.tsx
@@ -6,9 +6,9 @@ import { usePostDetails } from '@/hooks/usePostDetails/usePostDetails';
 import { parseCollectionContent } from '@/libs/post/collectionContent';
 import { isPostDeleted, isValidPostCompositeId } from '@/libs/utils/utils';
 import { parseCompositeId } from '@/models/models.utils';
-import { CollectionHero } from '@/organisms/CollectionHero/CollectionHero';
-import { CollectionItems } from '@/organisms/CollectionItems/CollectionItems';
-import { CollectionNotFound } from '@/organisms/CollectionNotFound/CollectionNotFound';
+import { CollectionHero } from '@/organisms/Collections/CollectionHero/CollectionHero';
+import { CollectionItems } from '@/organisms/Collections/CollectionItems/CollectionItems';
+import { CollectionNotFound } from '@/organisms/Collections/CollectionNotFound/CollectionNotFound';
 import { CollectionsSections } from '@/organisms/Collections/CollectionsSections/CollectionsSections';
 import { ContentLayout } from '@/organisms/ContentLayout/ContentLayout';
 import type { CollectionProps } from './Collection.types';
